backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0
python-decouple>=3.8
psycopg2-binary>=2.9
openpyxl>=3.1
gunicorn>=21.2
